+++ src/zope/component/zcml.py
@@ -304,6 +304,10 @@ def subscriber(_context, for_=None, factory=None, handler=None, provides=None,
         if handler is not None:
             raise TypeError("Cannot use handler with factory")
         if provides is None:
+            p = list(implementedBy(factory))
+            if len(p) == 1:
+                provides = p[0]
+        if provides is None:
             raise TypeError(
                 "You must specify a provided interface when registering "
                 "a factory")